5th test

+++ WeeklyTest_5/src/main/java/com/example/dataBase_Relationship/service/BookService.java
@@ -2,6 +2,8 @@ package com.example.dataBase_Relationship.service;
 
 import com.example.dataBase_Relationship.dao.BookRepo;
 import com.example.dataBase_Relationship.model.Book;
+import com.example.dataBase_Relationship.model.Student;
+import org.json.JSONArray;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,30 +13,42 @@ import java.util.List;
 public class BookService {
     @Autowired
     private BookRepo bookRepo;
-    public int addBook(Book book) {
+    public int saveBook(Book book) {
         Book bookObj=bookRepo.save(book);
         return bookObj.getBookId();
     }
 
-    public Book getBookById(Integer bookId) {
-        return bookRepo.findById(bookId).get();
+    public JSONArray getBook(String bookId) {
+        JSONArray bookArray=new JSONArray();
+        if(bookId!=null && bookRepo.findById(Integer.valueOf(bookId)).isPresent()){
+            Book book=bookRepo.findById(Integer.valueOf(bookId)).get();
+            bookArray.put(book);
+        }
+        else {
+            List<Book> booksList=bookRepo.findAll();
+            for(Book book:booksList){
+                bookArray.put(book);
+            }
+        }
+        return bookArray;
     }
-
-    public List<Book> getAllBooks() {
-        return bookRepo.findAll();
-    }
-
     public Book updateBook(Integer bookId, Book newBook) {
         Book book= bookRepo.findById(bookId).get();
         book.setAuthor(newBook.getAuthor());
         book.setBookTitle(newBook.getBookTitle());
-        book.setPrice(book.getPrice());
-        book.setDescription(book.getDescription());
-        book.setStudent(book.getStudent());
+        book.setPrice(newBook.getPrice());
+        book.setDescription(newBook.getDescription());
+        book.setStudent(newBook.getStudent());
         return bookRepo.save(book);
     }
 
-    public void deleteBook(Integer bookId) {
-         bookRepo.deleteById(bookId);
+    public boolean deleteBook(Integer bookId) {
+        if(bookRepo.findById(bookId).isPresent()){
+            bookRepo.deleteById(bookId);
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 }
